+++ src/Trans.js
@@ -11,6 +11,7 @@ function getChildren(node) {
 }
 
 function nodesToString(mem, children, index) {
+  if (!children) return '';
   if (Object.prototype.toString.call(children) !== '[object Array]') children = [children];
 
   children.forEach((child, i) => {
@@ -48,6 +49,7 @@ function nodesToString(mem, children, index) {
 
 function renderNodes(children, targetString, i18n) {
   if (targetString === "") return [];
+  if (!children) return [targetString];
 
   // parse ast from string with additional wrapper tag
   // -> avoids issues in parser removing prepending text nodes
@@ -97,7 +99,7 @@ export default class Trans extends React.Component {
 
   render() {
     const contextAndProps = { i18n: this.context.i18n, t: this.context.t, ...this.props };
-    const { children, count, parent, i18nKey, i18n, t: tFromContextAndProps, ...additionalProps } = contextAndProps;
+    const { children, count, parent, i18nKey, ns: namespace, i18n, t: tFromContextAndProps, ...additionalProps } = contextAndProps;
     const t = tFromContextAndProps || i18n.t.bind(i18n);
 
     const reactI18nextOptions = (i18n.options && i18n.options.react) || {};
@@ -106,7 +108,7 @@ export default class Trans extends React.Component {
     const defaultValue = nodesToString('', children, 0);
     const hashTransKey = reactI18nextOptions.hashTransKey;
     const key = i18nKey || (hashTransKey ? hashTransKey(defaultValue) : defaultValue);
-    const translation = key ? t(key, { interpolation: { prefix: '#$?', suffix: '?$#' }, defaultValue, count }) : defaultValue;
+    const translation = key ? t(key, { interpolation: { prefix: '#$?', suffix: '?$#' }, defaultValue, count, ns: namespace }) : defaultValue;
 
     if (reactI18nextOptions.exposeNamespace) {
       let ns = typeof t.ns === 'string' ? t.ns : t.ns[0];